wpt_babel/content-security-policy/style-src-attr-elem/style-src-attr-allowed-src-blocked.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="style-src-attr 'unsafe-inline';
    style-src 'none';" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script>
   "use strict";

var t = async_test("Should apply the style attribute");
window.addEventListener('securitypolicyviolation', t.unreached_func("Should not have fired a spv event"));
  </script>
 </head>
 <body style="background: green">
  <script>
   "use strict";

t.step(function () {
  assert_true(document.body.style.length > 0);
  t.done();
});
  </script>
 </body>
</html>
